// src/app/dometa-api/model/entityModel.ts
/**
 * DoMeta API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { DomainEventModel } from './domainEventModel';
import { EntityRelationModel } from './entityRelationModel';
import { PropertyModel } from './propertyModel';

export interface EntityModel { 
    id?: string;
    boundedContextId?: string;
    name?: string;
    identityPropertyName?: string;
    properties?: Array<PropertyModel>;
    relations?: Array<EntityRelationModel>;
    domainEvents?: Array<DomainEventModel>;
}