// src/app/dometa-api/model/valueObjectModel.ts
/**
 * DoMeta API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { PropertyModel } from './propertyModel';

export interface ValueObjectModel { 
    id?: string;
    boundedContextId?: string;
    name?: string;
    properties?: Array<PropertyModel>;
}